# SurveyTool/LearningPlatform/mongodb-api-documentation-swagger.yaml
...
swagger: '2.0'
info:
  version: v1
  title: ResponsiveInsight API
  description: |
    This API will be used both internally for the application, and externally for integration. 
    
    Each path below will have ETAGs. The /surveys/ and /surveydefintions/ will have different ETAGS. 
host: api-surveytool.orientsoftware.asia
basePath: /api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
...
parameters:
  surveyId:
    name: surveyId
    in: path
    description: The survey identifier
    type: string
    required: true
  folderId:
    name: folderId
    in: path
    description: The folder identifier
    type: string
    required: true
  pageId:
    name: pageId
    in: path
    description: The page identifier
    type: string
    required: true
  questionId:
    name: questionId
    in: path
    description: The question identifier
    type: string
    required: true
...
  reportId:
    name: reportId
    in: path
    description: the report identifier
    type: integer
    required: true
...
paths:
...
  /surveys/{surveyId}/folders/{folderId}/pages/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/folderId'
    post:
      tags: [page]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/folders/{folderId}/pages/{pageId}/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/folderId'
      - $ref: '#/parameters/pageId'
    put:
      tags: [page]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    delete:
      tags: [page]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/folders/{folderId}/pages/{pageId}/move/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/folderId'
      - $ref: '#/parameters/pageId'
    post:
      tags: [page]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...
  /surveys/{surveyId}/pages/{pageId}/questions/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/pageId'
    get:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    post:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/pages/{pageId}/questions/{questionId}/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/pageId'
      - $ref: '#/parameters/questionId'
    get:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    put:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    delete:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /surveys/{surveyId}/pages/{pageId}/questions/{questionId}/duplicate/:
    parameters:
      - $ref: '#/parameters/surveyId'
      - $ref: '#/parameters/pageId'
      - $ref: '#/parameters/questionId'
    post:
      tags: [question]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...
  /reports/{reportId}/definition/:
    parameters:
      - $ref: '#/parameters/reportId'
    get:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    put:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    post:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /reports/{reportId}/definition/pages/:
    parameters:
      - $ref: '#/parameters/reportId'
    get:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
    post:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /reports/{reportId}/definition/pages/{pageId}/:
    parameters:
      - $ref: '#/parameters/reportId'
      - $ref: '#/parameters/pageId'
    delete:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /reports/{reportId}/definition/pages/move/:
    parameters:
      - $ref: '#/parameters/reportId'
    patch:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
  /reports/{reportId}/definition/pages/questions/:
    parameters:
      - $ref: '#/parameters/reportId'
    get:
      tags: [reports]
      produces:
        - application/json
      responses:
        '200':
          description: OK
...